+++ test/base-crud.test.ts
@@ -1,6 +1,6 @@
 import { assert } from 'chai';
 import request from 'supertest';
-import { StatusCode } from '../src/StatusCode';
+import { StatusCode } from '../src/app/status-code';
 import { config } from "dotenv";
 
 // set env
@@ -37,7 +37,7 @@ describe('Start: Test base crud!', () => {
     it('responds with an empty array', (done) => {
       request(host)
         .get('/api/users')
-        .expect(StatusCode.Ok)
+        .expect(StatusCode.OK)
         .expect(JSON.stringify([]), done);
     });
   });
@@ -52,7 +52,7 @@ describe('Start: Test base crud!', () => {
       request(host)
         .post('/api/users')
         .send(JSON.stringify(testUser))
-        .expect(StatusCode.Created)
+        .expect(StatusCode.CREATED)
         .then(response => {
           const responseTestUser = JSON.parse(response.text);
           tempData.id = responseTestUser.id;
@@ -69,7 +69,7 @@ describe('Start: Test base crud!', () => {
     it('responds to user id', (done) => {
       request(host)
         .get(`/api/users/${tempData.id}`)
-        .expect(StatusCode.Ok)
+        .expect(StatusCode.OK)
         .then(response => {
           const responseTestUser = JSON.parse(response.text);
           assert.equal(responseTestUser.id, tempData.id);
@@ -89,7 +89,7 @@ describe('Start: Test base crud!', () => {
       request(host)
         .put(`/api/users/${tempData.id}`)
         .send(JSON.stringify(testUser))
-        .expect(StatusCode.Ok)
+        .expect(StatusCode.OK)
         .then(response => {
           const responseTestUser = JSON.parse(response.text);
           assert.equal(responseTestUser.id, tempData.id);
@@ -106,7 +106,7 @@ describe('Start: Test base crud!', () => {
     it('responds to user id and updates', (done) => {
       request(host)
         .delete(`/api/users/${tempData.id}`)
-        .expect(StatusCode.NoContent, done);
+        .expect(StatusCode.NO_CONTENT, done);
     });
   });
   
@@ -114,7 +114,7 @@ describe('Start: Test base crud!', () => {
     it('get a non-existent user', (done) => {
       request(host)
         .get(`/api/users/${tempData.id}`)
-        .expect(StatusCode.NotFound, done);
+        .expect(StatusCode.NOT_FOUND, done);
     });
   });
