+++ test/random-users-performance.test.ts
@@ -1,6 +1,6 @@
 import { assert } from 'chai';
 import request from 'supertest';
-import { StatusCode } from '../src/StatusCode';
+import { StatusCode } from '../src/app/status-code';
 import { config } from "dotenv";
 
 // set env
@@ -43,7 +43,7 @@ describe('Start: Test more users crud!', () => {
     it('responds with an empty array', (done) => {
       request(host)
         .get('/api/users')
-        .expect(StatusCode.Ok)
+        .expect(StatusCode.OK)
         .expect(JSON.stringify([]), done);
     });
   });
@@ -55,7 +55,7 @@ describe('Start: Test more users crud!', () => {
         request(host)
           .post('/api/users')
           .send(JSON.stringify(randomUser))
-          .expect(StatusCode.Created)
+          .expect(StatusCode.CREATED)
           .then(response => {
             const responseTestUser = JSON.parse(response.text);
             tempDataId.push(responseTestUser.id);
@@ -74,7 +74,7 @@ describe('Start: Test more users crud!', () => {
       it(`User ${index + 1} is exist...`, (done) => {
         request(host)
           .get(`/api/users/${tempDataId[index]}`)
-          .expect(StatusCode.Ok)
+          .expect(StatusCode.OK)
           .then(response => {
             const responseTestUser = JSON.parse(response.text);
             assert.equal(responseTestUser.id, tempDataId[index]);
@@ -90,7 +90,7 @@ describe('Start: Test more users crud!', () => {
       it(`Delete user ${index + 1}...`, (done) => {
         request(host)
           .delete(`/api/users/${tempDataId[index]}`)
-          .expect(StatusCode.NoContent, done);
+          .expect(StatusCode.NO_CONTENT, done);
       });
     }
   });
@@ -100,7 +100,7 @@ describe('Start: Test more users crud!', () => {
       it(`User ${index + 1} deleted...`, (done) => {
         request(host)
           .delete(`/api/users/${tempDataId[index]}`)
-          .expect(StatusCode.NotFound, done);
+          .expect(StatusCode.NOT_FOUND, done);
       });
     }
   });
